requirements.txt
langchain==0.1.0
sentence-transformers==2.2.2
chromadb==0.4.18
openai==1.6.1
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
streamlit==1.29.0
scikit-learn==1.3.2